change slice dims to int32 from int64

+++ lite/backends/nnadapter/nnadapter/core/operation/slice.cc
@@ -26,7 +26,7 @@ namespace nnadapter {
 namespace operation {
 
 static void SliceOfflineCalc(const int32_t* input,
-                             std::vector<int64_t> in_dims,
+                             std::vector<int32_t> in_dims,
                              uint32_t axes_count,
                              int32_t* axes,
                              int32_t* starts,
@@ -112,21 +112,23 @@ int PrepareSlice(hal::Operation* operation) {
     NNAdapterOperandDimensionType dimension_type;
     dimension_type.count = output_operand->type.dimensions.data[0];
     dimension_type.dynamic_count = input_operand->type.dimensions.dynamic_count;
-    SliceOfflineCalc(tempory_shape_info.data,
-                     std::vector<int64_t>({dimension_type.count}),
-                     axes_count,
-                     axes,
-                     starts,
-                     ends,
-                     dimension_type.data);
+    SliceOfflineCalc(
+        tempory_shape_info.data,
+        std::vector<int32_t>({static_cast<int32_t>(dimension_type.count)}),
+        axes_count,
+        axes,
+        starts,
+        ends,
+        dimension_type.data);
     for (uint32_t i = 0; i < dimension_type.dynamic_count; i++) {
-      SliceOfflineCalc(tempory_shape_info.dynamic_data[i],
-                       std::vector<int64_t>({dimension_type.count}),
-                       axes_count,
-                       axes,
-                       starts,
-                       ends,
-                       dimension_type.dynamic_data[i]);
+      SliceOfflineCalc(
+          tempory_shape_info.dynamic_data[i],
+          std::vector<int32_t>({static_cast<int32_t>(dimension_type.count)}),
+          axes_count,
+          axes,
+          starts,
+          ends,
+          dimension_type.dynamic_data[i]);
     }
     output_operand->hints[NNADAPTER_TEMPORY_SHAPE_INFO].set(dimension_type);
   }
